Add files via upload
This is my first HTML and JavaScript project.  This game is modeled after cookie clicker and is still in its beginning stages.  The egg and dragon pictures I created myself, but the rest of the images are from the internet.  The ascend function does not work and the button is just there for looks.  I'll continue to work on the local files on my computer and upload new versions.